# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Mobile subscription and referral API for Mama Bhutnika
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MobileSignupInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  referralCode: string;
}

export interface MobileLoginInput {
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface ReferralInput {
  /** @minLength 1 */
  referralCode: string;
}

export interface GrantDaysInput {
  /**
   * @minimum 1
   * @maximum 3650
   */
  days: number;
}

export interface User {
  uid: string;
  name: string;
  email: string;
  referralCode: string;
  subscriptionUntil: string;
  subscriptionActive: boolean;
  rewardProgress: number;
  totalAdsWatched: number;
  isAdmin: boolean;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UserResponse {
  user: User;
}

export interface AdminUsersResponse {
  users: User[];
}

export interface MessageResponse {
  message: string;
}

/**
 * Invalid request
 */
export type BadRequestResponse = ErrorResponse;

/**
 * Authentication required
 */
export type UnauthorizedResponse = ErrorResponse;

/**
 * Administrator access required
 */
export type ForbiddenResponse = ErrorResponse;

/**
 * Resource conflict
 */
export type ConflictResponse = ErrorResponse;

/**
 * Resource not found
 */
export type NotFoundResponse = ErrorResponse;