# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mobile subscription and referral API for Mama Bhutnika
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: mobile
    description: Mobile app operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /mobile/auth/signup:
    post:
      operationId: mobileSignup
      tags: [mobile]
      summary: Create a mobile account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MobileSignupInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /mobile/auth/login:
    post:
      operationId: mobileLogin
      tags: [mobile]
      summary: Sign in to a mobile account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MobileLoginInput"
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /mobile/me:
    get:
      operationId: mobileMe
      tags: [mobile]
      summary: Get the current mobile account
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /mobile/referrals/redeem:
    post:
      operationId: mobileRedeemReferral
      tags: [mobile]
      summary: Redeem a referral code
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReferralInput"
      responses:
        "200":
          description: Referral redeemed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
  /mobile/rewards/ad-completed:
    post:
      operationId: mobileRewardAdCompleted
      tags: [mobile]
      summary: Confirm a completed rewarded ad
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /mobile/admin/users:
    get:
      operationId: mobileAdminUsers
      tags: [mobile]
      summary: List mobile accounts for an administrator
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Account list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUsersResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /mobile/admin/users/{uid}/grant:
    post:
      operationId: mobileAdminGrantDays
      tags: [mobile]
      summary: Grant subscription days to an account
      security:
        - bearerAuth: []
      parameters:
        - name: uid
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GrantDaysInput"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Administrator access required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MobileSignupInput:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
        password:
          type: string
          minLength: 6
        referralCode:
          type: string
      required: [name, email, password, referralCode]
    MobileLoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 6
      required: [email, password]
    ReferralInput:
      type: object
      properties:
        referralCode:
          type: string
          minLength: 1
      required: [referralCode]
    GrantDaysInput:
      type: object
      properties:
        days:
          type: number
          minimum: 1
          maximum: 3650
      required: [days]
    User:
      type: object
      properties:
        uid:
          type: string
        name:
          type: string
        email:
          type: string
        referralCode:
          type: string
        subscriptionUntil:
          type: string
          format: date-time
        subscriptionActive:
          type: boolean
        rewardProgress:
          type: number
        totalAdsWatched:
          type: number
        isAdmin:
          type: boolean
      required:
        - uid
        - name
        - email
        - referralCode
        - subscriptionUntil
        - subscriptionActive
        - rewardProgress
        - totalAdsWatched
        - isAdmin
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
    UserResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
      required: [user]
    AdminUsersResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
      required: [users]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]