# lib/api-zod/src/generated/types/adminUsersResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Mobile subscription and referral API for Mama Bhutnika
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export interface AdminUsersResponse {
  users: User[];
}