# lib/api-zod/src/generated/types/mobileSignupInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Mobile subscription and referral API for Mama Bhutnika
 * OpenAPI spec version: 0.1.0
 */

export interface MobileSignupInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  referralCode: string;
}